ajout ennemi cracheur de feu et déplacement fluide des ennemis

+++ src/Character.h
@@ -7,13 +7,13 @@ class Character : public Element
 public:
 	Character(double x, double y);
 	int get_speed() { return speed; }
-	virtual void move(int x, int y) = 0;
+	virtual void move(double x, double y) = 0;
 	virtual void attack() = 0;
 
 
 
 protected:
-	int speed = 3;
+	double speed = 3;
 	int maxHP;
 	int currentHP;
 };
