+++ src/Commandes.h
@@ -4,15 +4,15 @@
 class Command {
 public:
 	virtual ~Command() {}
-	virtual void execute(Character& chara, int speed) = 0;
+	virtual void execute(Character& chara, double speed) = 0;
 };
 class MoveUpCommand : public Command {
 	public:
 		MoveUpCommand() {
 
 		};
-	virtual void execute(Character& chara, int speed) { moveUp(chara, speed); }
-	void moveUp(Character& chara, int speed) {
+	virtual void execute(Character& chara, double speed) { moveUp(chara, speed); }
+	void moveUp(Character& chara, double speed) {
 		chara.move(0, -speed);
 	};
 
@@ -21,8 +21,8 @@ class MoveUpCommand : public Command {
 class MoveDownCommand : public Command {
 public:
 	MoveDownCommand() {};
-	virtual void execute(Character& chara, int speed) { moveDown(chara, speed); }
-	void moveDown(Character& chara, int speed) {
+	virtual void execute(Character& chara, double speed) { moveDown(chara, speed); }
+	void moveDown(Character& chara, double speed) {
 		chara.move(0,speed);
 	};
 };
@@ -30,8 +30,8 @@ public:
 class MoveRightCommand : public Command {
 public:
 	MoveRightCommand() {};
-	virtual void execute(Character& chara, int speed) { moveRight(chara, speed); }
-	void moveRight(Character& chara, int speed) {
+	virtual void execute(Character& chara, double speed) { moveRight(chara, speed); }
+	void moveRight(Character& chara, double speed) {
 		chara.move(speed, 0);
 	};
 };
@@ -39,8 +39,8 @@ public:
 class MoveLeftCommand : public Command {
 public:
 	MoveLeftCommand() {};
-	virtual void execute(Character& chara, int speed) { moveLeft(chara, speed); }
-	void moveLeft(Character& chara, int speed) {
+	virtual void execute(Character& chara, double speed) { moveLeft(chara, speed); }
+	void moveLeft(Character& chara, double speed) {
 		chara.move(-speed, 0);
 	};
 
@@ -48,7 +48,7 @@ public:
 class AttackCommand : public Command {
 public:
 	AttackCommand() {};
-	virtual void execute(Character& chara, int speed) { attack(chara); }
+	virtual void execute(Character& chara, double speed) { attack(chara); }
 	void attack(Character& chara) {
 
 		chara.attack();
@@ -58,6 +58,6 @@ public:
 class PowerUpCommand : public Command {
 public:
 	PowerUpCommand() {};
-	virtual void execute(Character& chara, int speed) { powerup(chara); }
+	virtual void execute(Character& chara, double speed) { powerup(chara); }
 	void powerup(Character& chara) {};
 };
