+++ src/Hero.h
@@ -16,7 +16,7 @@ class Hero : public Character
 public:
 	Hero(double x_init, double y_init, double size_init);
 	void handle_keyboard();
-	void move(int x, int y);
+	void move(double x, double y);
 	void update_sprite();
 	void attack();
